refactor: 리뷰 반영  - 클래스명 변경 및 패키지 경로 변경  - 코드 간소화
