+++ src/main/java/lotto/view/ResultView.java
@@ -3,6 +3,7 @@ package lotto.view;
 import lotto.domain.Lotto;
 import lotto.domain.Lottos;
 import lotto.domain.Rank;
+import lotto.domain.Result;
 
 import java.util.Map;
 
@@ -13,13 +14,10 @@ public class ResultView {
     private static final String RESULT_FORMAT = "총 수익률은 %.2f 입니다.";
     private static final String LOSS_MESSAGE = "(기준이 1이기 때문에 결과적으로 손해라는 의미임)";
 
-    private ResultInfo resultInfo;
-
-    public ResultView() {
-    }
+    private final Result result;
 
     public ResultView(Lottos lottos, Lotto answer) {
-        this.resultInfo = new ResultInfo(lottos, answer);
+        this.result = new Result(lottos, answer);
     }
 
     public static void printLottos(Lottos lottos) {
@@ -29,14 +27,14 @@ public class ResultView {
     }
 
     public void printResult() {
-        Map<Integer, Integer> result = resultInfo.getResultInfo();
+        Map<Integer, Integer> result = this.result.getResultInfo();
         System.out.println(STATISTIC);
         for (Rank rank : Rank.values()) {
             System.out.println(String.format(MATCH_FORMAT, rank.getCount(), rank.getAmount(), result.get(rank.getCount())));
         }
 
-        System.out.print(String.format(RESULT_FORMAT, resultInfo.getRateOfReturn()));
-        if (resultInfo.getRateOfReturn() < 1) {
+        System.out.print(String.format(RESULT_FORMAT, this.result.getRateOfReturn()));
+        if (this.result.getRateOfReturn() < 1) {
             System.out.println(LOSS_MESSAGE);
         }
     }
